# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateWorkspaceRequest": {
        "properties": {
          "branch": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "location": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Location",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/RepositorySource"
          },
          "token": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          }
        },
        "required": [
          "name",
          "source",
          "location"
        ],
        "title": "CreateWorkspaceRequest",
        "type": "object"
      },
      "FileManifestEntryResponse": {
        "properties": {
          "indexable": {
            "title": "Indexable",
            "type": "boolean"
          },
          "is_binary": {
            "title": "Is Binary",
            "type": "boolean"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "path",
          "sha256",
          "size_bytes",
          "is_binary",
          "indexable"
        ],
        "title": "FileManifestEntryResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IndexStateResponse": {
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "file_count": {
            "title": "File Count",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/IndexStatus"
          },
          "symbol_count": {
            "title": "Symbol Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "status",
          "file_count",
          "symbol_count",
          "chunk_count",
          "updated_at"
        ],
        "title": "IndexStateResponse",
        "type": "object"
      },
      "IndexStatus": {
        "enum": [
          "pending",
          "building",
          "ready",
          "failed"
        ],
        "title": "IndexStatus",
        "type": "string"
      },
      "Language": {
        "enum": [
          "python",
          "javascript",
          "typescript",
          "go",
          "java",
          "rust"
        ],
        "title": "Language",
        "type": "string"
      },
...
      "RepositorySource": {
        "enum": [
          "local",
          "github"
        ],
        "title": "RepositorySource",
        "type": "string"
      },
...
      "SearchHitResponse": {
        "description": "One hybrid-search result with full provenance (FR-2.3).\n\n``suspect`` surfaces the index-time injection screen so a client renders a\nwarning; ``source`` reports how the hit was found ('hybrid' or 'symbol').",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "end_line": {
            "title": "End Line",
            "type": "integer"
          },
          "header_path": {
            "title": "Header Path",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/SymbolKind"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "start_line": {
            "title": "Start Line",
            "type": "integer"
          },
          "suspect": {
            "title": "Suspect",
            "type": "boolean"
          }
        },
        "required": [
          "path",
          "language",
          "header_path",
          "kind",
          "start_line",
          "end_line",
          "content",
          "score",
          "suspect",
          "source"
        ],
        "title": "SearchHitResponse",
        "type": "object"
      },
      "SearchResponse": {
        "properties": {
          "hits": {
            "items": {
              "$ref": "#/components/schemas/SearchHitResponse"
            },
            "title": "Hits",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query",
          "hits"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "SymbolKind": {
        "description": "Language-neutral symbol categories. Each language maps its native node\ntypes onto this set (docs/06); `import` captures every dependency form.",
        "enum": [
          "class",
          "function",
          "method",
          "interface",
          "struct",
          "enum",
          "trait",
          "import"
        ],
        "title": "SymbolKind",
        "type": "string"
      },
      "SymbolResponse": {
        "properties": {
          "end_line": {
            "title": "End Line",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/SymbolKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          },
          "qualified_name": {
            "title": "Qualified Name",
            "type": "string"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "start_line": {
            "title": "Start Line",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "name",
          "qualified_name",
          "parent",
          "start_line",
          "end_line",
          "reference"
        ],
        "title": "SymbolResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkspaceResponse": {
        "properties": {
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "file_count": {
            "title": "File Count",
            "type": "integer"
          },
          "head_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Commit"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/RepositorySource"
          },
          "status": {
            "$ref": "#/components/schemas/WorkspaceStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "owner_id",
          "name",
          "source",
          "location",
          "branch",
          "status",
          "head_commit",
          "size_bytes",
          "file_count",
          "error",
          "created_at",
          "updated_at"
        ],
        "title": "WorkspaceResponse",
        "type": "object"
      },
      "WorkspaceStatus": {
        "enum": [
          "pending",
          "ingesting",
          "ready",
          "failed"
        ],
        "title": "WorkspaceStatus",
        "type": "string"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Autonomous coding agent platform",
    "title": "Spidey API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces": {
      "get": {
        "operationId": "list_workspaces_api_v1_workspaces_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceResponse"
                  },
                  "title": "Response List Workspaces Api V1 Workspaces Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List your workspaces",
        "tags": [
          "workspaces"
        ]
      },
      "post": {
        "operationId": "create_workspace_api_v1_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a workspace and start ingestion",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}": {
      "delete": {
        "operationId": "delete_workspace_api_v1_workspaces__workspace_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete a workspace and its files",
        "tags": [
          "workspaces"
        ]
      },
      "get": {
        "operationId": "get_workspace_api_v1_workspaces__workspace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get a workspace",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/files": {
      "get": {
        "operationId": "list_workspace_files_api_v1_workspaces__workspace_id__files_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FileManifestEntryResponse"
                  },
                  "title": "Response List Workspace Files Api V1 Workspaces  Workspace Id  Files Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the ingested file manifest",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/index": {
      "get": {
        "operationId": "get_index_state_api_v1_workspaces__workspace_id__index_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Code-index status for a workspace",
        "tags": [
          "workspaces"
        ]
      },
      "post": {
        "operationId": "reindex_code_api_v1_workspaces__workspace_id__index_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Re-run code indexing for a workspace",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/ingest": {
      "post": {
        "operationId": "reingest_workspace_api_v1_workspaces__workspace_id__ingest_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Re-run ingestion for a workspace",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/search": {
      "get": {
        "operationId": "search_code_api_v1_workspaces__workspace_id__search_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "description": "Search query",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Search query",
              "maxLength": 1024,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Hybrid (semantic + lexical) code search",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/symbols": {
      "get": {
        "operationId": "list_symbols_api_v1_workspaces__workspace_id__symbols_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SymbolResponse"
                  },
                  "title": "Response List Symbols Api V1 Workspaces  Workspace Id  Symbols Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Extracted symbols for a workspace",
        "tags": [
          "workspaces"
        ]
      }
    }
  }
}